pyproject.toml
[tool.poetry]
name = "tradingview"
version = "3.0.0-rc.1"
description = "A package for creating stock screeners with the TradingView API"
authors = ["shner-elmo <770elmo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tradingview", from = "src"}]
repository = "https://github.com/shner-elmo/TradingView-Screener"
documentation = "https://github.com/shner-elmo/TradingView-Screener/blob/master/README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Environment :: Console",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = ">=3.9"
requests = "2.*"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
pandas = "^2.2.3"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.156.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
ruff = "^0.1.3"
pdoc = "^15.0.0"
pytest = "^7.4.3"
pyright = "^1.1.364"

#[tool.poetry.extras]
#df = ["pandas"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint.pycodestyle]
max-doc-length = 88
#max-line-length = 100

[tool.ruff.format]
quote-style = "single"

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.flake8-errmsg]
max-string-length = 20

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.9"
include = [
    "src/tradingview_screener/**",
    "tests",
]
